requirements.txt
feedparser==5.2.1
python-dateutil==2.4.2
marshmallow==2.0.0b3
django-registration==1.0
html2text==2015.11.4
Markdown==2.6.4
cached_property==1.3.0
geopy==1.21.0
google_api_python_client==1.6.4
xlrd==1.2.0
